requirements.txt
progressbar2
PyObjC
pydub 
simpleaudio